refactor: 기존 util 도메인 내부 패키지로 이동
